+++ src/pages/AssetClass.vue
@@ -33,6 +33,7 @@
               type="textarea"
               rows="3"
               :max-height="300"
+              autofocus @keyup.enter.stop
             />
           </q-popup-edit>
         </q-card-section>