+++ src/pages/Mission.vue
@@ -71,12 +71,14 @@
 
         <q-separator/>
         <q-card-section>
+          Description:
           <mxm-markdown :text="mission.description"/>
           <q-popup-edit
             v-if="mission.missionStatus === 'DRAFT'"
             v-model="mission.description"
             title="Description"
             buttons
+            persistent
             @save="updateDescription"
           >
             <q-input
@@ -86,6 +88,7 @@
               type="textarea"
               rows="3"
               :max-height="300"
+              autofocus @keyup.enter.stop
             />
           </q-popup-edit>
         </q-card-section>