+++ src/pages/Parameter.vue
@@ -105,6 +105,7 @@
                     type="textarea"
                     rows="3"
                     :max-height="300"
+                    autofocus @keyup.enter.stop
                   />
                 </q-popup-edit>
               </td>