+++ ChangeLog.md
@@ -1,5 +1,8 @@
 2019-11
 
+- use @keyup.enter.stop on multiline inputs under q-popup-edit.
+  https://quasar.dev/vue-components/popup-edit#Textarea-%2F-QEditor
+
 - starting to use qgeomap
 - upgrading to Quasar 1
 - factoring out and extending geometry viewing/editing into qgeomap, a separate quasar extension